change __init__
